#430: the empire's maintenance total is a RECEIVER SLOT, not a cache beside the package
A cross-scope receiver total is the sum of its members' realized values, and the empire's
maintenance is exactly that sum over its cities -- so it belongs in the same cache holding a
different slot, never in a hand-named CvDerivedCache of its own. A named field cannot be
addressed by a derived mask, which is why it forces a bespoke invalidation path and why
DEC-uniform-cache-shape calls the shape a defect rather than untidiness.

Maintenance is the first NON-commerce receiver, which is what makes the rule general instead of
a commerce habit. Two things fall out of that:

  - its per-city quantity is the one a package cannot answer alone (the city's realized value
    folds the four engine components and declines wholesale under WLTKD/disorder), so the sum
    asks the city for its realized value, and the oracle recomputes BOTH halves from source
    rather than inheriting either off the stored surface;

  - the cross-scope roll-up answers a receiver channel with its maintained sum, so CvPlayer's
    percent stack can no longer go through it -- routing it there would have handed back the
    realized TOTAL where a percent stack was asked for, silently and plausibly. It reads the
    rolled legs directly now.

The receiver bit region moves 59 -> 58 to hold five slots; bit 63 stays the tripwire.

Two width defects go with it, both the shape fixed-point-and-scales.md 1b names:

  - the empire total accumulated into a `long`, which on VC7.1 is 32 bits -- it reads as
    deliberate headroom and provides none, then narrowed again into an int cache, mid-way
    through a money plane that is already int64_t;

  - every empire receiver sum was cast (int) into its int64 slot on both the rebuild and the
    oracle path. That one is not maintenance-specific: it was narrowing culture, the plane the
    late-game overflow was found on.

Readers narrow at their own boundary, which is where the out-boundary conversion belongs.

+++ Sources/AI/CvCityAI.cpp
@@ -5644,7 +5644,7 @@ int CvCityAI::AI_buildingValueThresholdOriginalUncached(BuildingTypes eBuilding,
 			if (((iFocusFlags & BUILDINGFOCUS_MAINTENANCE) || (iFocusFlags & BUILDINGFOCUS_GOLD) || iPass > 0)
 			&& kBuilding.getCommerceChange(COMMERCE_GOLD) < 0 && GC.getTREAT_NEGATIVE_GOLD_AS_MAINTENANCE())
 			{
-				const int iBaseMaintenance = getMaintenanceTimes100();
+				const int iBaseMaintenance = (int)getMaintenanceTimes100();
 				const int iMaintenanceMod = maintenancePercentStack((int)MAINTENANCE_AMOUNT);
 
 				const int iCost =
